+++ package/views/index.blade.php
@@ -21,7 +21,8 @@
 	<script id="admin" type="text/plain">{!! json_encode($admin, JSON_UNESCAPED_SLASHES) !!}</script>
 	<script id="settings" type="text/plain">{!! json_encode($settings, JSON_UNESCAPED_SLASHES) !!}</script>
 	<script id="data" type="text/plain">{!! json_encode($data, JSON_UNESCAPED_SLASHES) !!}</script>
-	<script id="home" type="text/template">{!! $home !!}</script>
+	<script id="page-home" type="text/template">{!! $home !!}</script>
+	<script id="page-help" type="text/template">{!! $help !!}</script>
 
     <!-- app -->
     <script src="{{ $assets }}js/app.js"></script>