+++ package/views/no-help.blade.php
@@ -1,0 +1,14 @@
+<header>
+	<h1>Help</h1>
+</header>
+<section>
+	<h3>Error</h3>
+	<p>Custom help page <code>{{ $help }}</code> not found.</p>
+	<p>You can:</p>
+	<ul>
+		<li>create it manually</li>
+		<li>re-run setup</li>
+		<li>switch back to a default help page</li>
+		<li>choose not show a help page</li>
+	</ul>
+</section>