+++ src/controllers/ApiController.php
@@ -82,6 +82,17 @@ class ApiController extends Controller
 		}
 
 		/**
+		 * Loads custom page content
+		 *
+		 * @param   string  $name
+		 * @return  string
+		 */
+		public function page($name)
+		{
+			return view("sketchpad::$name");
+		}
+
+		/**
 		 * Loads or saves settings data
 		 *
 		 * @method  POST
@@ -92,6 +103,11 @@ class ApiController extends Controller
 		 */
 		public function settings(Request $request, SketchpadConfig $config)
 		{
+			if (!$config->admin->settings)
+			{
+				return response()->json($config->settings);
+			}
+
 			function textToArray ($text)
 			{
 				return array_values(array_filter(array_map('trim', explode("\n", trim($text))), 'strlen'));