+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## [1.1.6] - 2017-05-03
+## [1.1.7] - 2017-05-05
 
 ### Added
 
@@ -13,19 +13,28 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Implemented controller reordering
 - Implemented support for implicit pagination
 - New `text()` and `code()` helper functions
+- Data injection for `md()` function
 - Static `Sketchpad::$form` and `Sketchpad::$data` properties
 - Automatic injection of `$route` variable into views
 - Included app controllers folder in paths
 - Custom page title functionality
-- Index pages for all help controllers
+- Added custom Help page
+- Implemented live-reloading for custom pages
+- Added scroll to section functionality for Settings
+- Added index pages for all help controllers
 - Bootstrap formatting for Markdown tables
 - Added changelog
 
 ### Changed
 
-- Move various settings to single "Site" section
+- Moved various settings to single "Site" section
+- New Param sub-components for different data types
+- Disabled Settings page now shows a message
 - Improved example controller and view content
+- Created new Sketchpad docs sections
 - Updated a lot of help to be better-formatted and more consistent
+- Index pages can be re-run
+- Changed some site key names
 - Added formatting classes to `pr()` and `vd()`
 - Updated FontAwesome to latest version
 - Simplified copying of user views in setup
@@ -35,6 +44,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - Defended against showing parent paths in Browse filesystem
+- Defended against saving settings if not allowed in admin.json
 - Fixed edge-case with buggy internal routes
 - Fixed bug with method titles showing parameters
 - Fixed Markdown not showing code or blockquotes correctly
@@ -42,8 +52,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fixed scroll to top animation
 - Fixed bug with intercepted link currentTarget
 - Fixed URL encoding bug with route parameters
+- Fixed bug in NumberParam where unset variables don't show 0
+- Fixed bug where Output only styles first markdown table
+- Fixed bug in internal href links not working for links with hashes
 - Removed absolute path references from RouteReference classes
 - Updated demo link in readme
+
+### Upgrade notes
+
+New Settings `site` section:
+
+- Copy `site` node from `vendor/davestewart/sketchpad/package/config/admin.json`
 
 
 ## [1.1.0] - 2017-04-21